added buyer backend unit tests

+++ backend/routes/buyer.js
@@ -62,7 +62,6 @@ router.post('/removeCartRegItem/:id', async (req, res) => {
 router.get('/cart', async (req, res, next) => {
   try {
     const user = await User.findOne({ email: req.session.email });
-    console.log(user);
     const cart = await ItemRegular.find({ _id: { $in: user.shoppingCart } });
     res.status(200).json(cart);
   } catch (error) {
@@ -80,8 +79,7 @@ router.post('/addWatchRegItem/:id', async (req, res) => {
       { $addToSet: { watchlistRegular: item }});
     res.status(200).send('Regular listing successfully added to watchlist!');
   } catch (error) {
-    res.status(500).send('An unknown error occured');
-    throw new Error('Error with adding reg item to watchlist');
+    res.status(500).send('Error with adding reg item to watchlist');
   }
 });
 
@@ -95,11 +93,11 @@ router.post('/addWatchBidItem/:id', async (req, res) => {
       { $addToSet: { watchlistBid: item }});
     res.status(200).send('Bid listing successfully added to watchlist!');
   } catch (error) {
-    res.status(500).send('An unknown error occured');
-    throw new Error('Error with adding bid item to watchlist');
+    res.status(500).send('Error with adding bid item to watchlist');
   }
 });
 
+/*
 // route to remove reg item from watchlist
 router.post('/removeWatchRegItem/:id', async (req, res) => {
   try {
@@ -115,7 +113,7 @@ router.post('/removeWatchRegItem/:id', async (req, res) => {
 });
 
 // route to remove bid item from watchlist
-router.post('/removeWatchRegItem/:id', async (req, res) => {
+router.post('/removeWatchBidItem/:id', async (req, res) => {
   try {
     const item = await ItemBid.findById(req.params.id);
     await User.findOneAndUpdate(
@@ -127,6 +125,7 @@ router.post('/removeWatchRegItem/:id', async (req, res) => {
     throw new Error('Error with removing bid item from watchlist');
   }
 });
+*/
 
 // route to add bid to bidhistory
 router.post('/addBid/:id', async (req, res) => {
@@ -140,8 +139,7 @@ router.post('/addBid/:id', async (req, res) => {
     );
     res.status(200).send('Bid placed successfully');
   } catch (error) {
-    res.status(500).send('An unknown error occured');
-    throw new Error('Error with adding bid to item');
+    res.status(500).send('Error with adding bid to item');
   }
 });
 
@@ -162,7 +160,7 @@ router.post('/regTransaction', async (req, res) => {
     });
     res.status(201).json(transaction);
   } catch (error) {
-    throw new Error('Error with completing transaction');
+    res.status(500).send('Error with completing Transaction');
   }
 });
 
@@ -200,10 +198,12 @@ router.post('/addTransaction', async (req, res) => {
       { name: transaction.seller },
       { $addToSet: { transactionHistory: transaction } },
     );
-    await ItemRegular.findByIdAndDelete(transaction.listingRegular._id);
-    res.status(200).send('Regular listing successfully added to watchlist!');
+    if (transaction.listingRegular) {
+      await ItemRegular.findByIdAndDelete(transaction.listingRegular._id);
+    }
+    res.status(200).send('Transaction successfully processed');
   } catch (error) {
-    throw new Error('Error with completeing transaction');
+    res.status(500).send('Error with completing Transaction');
   }
 });
